requirements.txt
keras==2.4.3
tensorflow-gpu==2.3.1
sentence-transformers==0.3.9
pandas